Always update all player actions
This fixes a bug: only the first player that changed (or was added) in a single
main context iteration was updated in update_player_action().  That's because
only one source id was used for all idles.

Another fix to this would be to keep track of which actions need to be updated.
This patch is a bit more brute-force, but that will probably not matter much,
as the amount of players is generally small.