Revert "Merge pull request #19 from mateosalta/patch-1"
This attempt to fix the microphone icon in the sound indicator doesn't
fix the problem and breaks the test. I decided to revert it so that
when the correct fix is in place (which is outside the indicator), it'll
work as intended.

This reverts commit d73cef63a75cb313842436266f5158de6ac3fe07, reversing
changes made to 9dd32c3ac9392a5412d8274cb198e0d6687e982c.