src/sound-menu.vala: Do not force panel icons in the sound menu
Panel icons are for the panel (mostly symbolic and monochromatic). Somewhere along the way, the Ayatana Sound Indicator menu was forced to use the same icons as those that were meant for the panel. This change makes the indicator menu use the same icons as any other menu or GUI element on the system.

End result: makes any theme capable of becoming Ayatana-friendly without hacks.